+++ Changelog.md
@@ -34,4 +34,7 @@
 ## 30/3/25 (Pranav)
 - added source code link in footer
 - some text changes on home page
-- added gradient week events, changed the design of that
+- added gradient week events, changed the design of that page
+
+## 31/3/25 (Pranav)
+- Added cultural evening performers